blog_garyrong/requirements.txt
django==1.6.5
-e git+https://github.com/duoshuo/duoshuo-python-sdk.git#egg=duoshuo-python-sdk
django-pingback
django-xmlrpc==0.1.4
gunicorn==18.0
BeautifulSoup==3.2.1
supervisor==3.0
Pygments==1.6
python-memcached==1.53
django-debug-toolbar==0.10.2
django_reversion==1.8.4
django_xadmin==0.5.0
django-ipware
django-reversion
django-markdown-deux
